requirements.txt
crewai
python-dotenv 
pandas 
numpy 
matplotlib 
seaborn
scikit-learn
streamlit 
plotly
google.generativeai